util: update zeta-switch()

+++ 3rd/zeta-switch.xsh
@@ -19,7 +19,7 @@ fi
 # 终端显示 NodeJS 安装位置 $ npm config get prefix
 if @zeta:xsh:has-cmd node; then
   NODE_PATH="$(command -v node)"
-  NODE_PATH="$(realpath "${NODE_PATH}")"
+  NODE_PATH="$(realpath -eq "${NODE_PATH}")"
   export NODE_PATH="${NODE_PATH%/bin/node}/lib/node_modules"
 fi
 # 中科 https://mirrors.ustc.edu.cn/help/node.html
@@ -27,9 +27,9 @@ fi
 
 # 终端显示当前 Go 环境变量 $ go env
 if @zeta:xsh:has-cmd go; then
-  GOROOT="$(command -v go)"         # Go 语言的安装目录
-  GOROOT="$(realpath "${GOROOT}")"  # => go/X.Y.Z/bin/go
-  export GOROOT="${GOROOT%/bin/go}" # => go/X.Y.Z
+  GOROOT="$(command -v go)" # Go 语言的安装目录
+  GOROOT="$(realpath -eq "${GOROOT}")" # => go/X.Y.Z/bin/go
+  export GOROOT="${GOROOT%/bin/go}"    # => go/X.Y.Z
 
   [[ "${GOROOT}" =~ "^${ZETA_DIR}/3rd/vendor/go/[0-9.]*$" ]] && {
     export GO111MODULE=on # 启用 module-aware 模式
@@ -58,16 +58,16 @@ function zman() {
   local cmakeMAN  nodeMAN  rustMAN  javaMAN  xMAN
 
   cmakeMAN="$(@zeta:3rd:get-vendor-path cmake)"
-  [[ -n "${cmakeMAN}" ]] && cmakeMAN="$(realpath "${cmakeMAN}/../man")"
+  [[ -n "${cmakeMAN}" ]] && cmakeMAN="$(realpath -eq "${cmakeMAN}/../man")"
 
   nodeMAN="$(@zeta:3rd:get-vendor-path node)"
-  [[ -n "${nodeMAN}" ]] && nodeMAN="$(realpath "${nodeMAN}/../share/man")"
+  [[ -n "${nodeMAN}" ]] && nodeMAN="$(realpath -eq "${nodeMAN}/../share/man")"
 
   javaMAN="$(@zeta:3rd:get-vendor-path java)"
-  [[ -n "${javaMAN}" ]] && javaMAN="$(realpath "${javaMAN}/../man")"
+  [[ -n "${javaMAN}" ]] && javaMAN="$(realpath -eq "${javaMAN}/../man")"
 
   rustMAN="$(@zeta:3rd:get-vendor-path rustc)"
-  [[ -n "${rustMAN}" ]] && rustMAN="$(realpath "${rustMAN}/../share/man")"
+  [[ -n "${rustMAN}" ]] && rustMAN="$(realpath -eq "${rustMAN}/../share/man")"
 
   if [[ $# -eq 3 ]]; then
     case "$1" in
@@ -121,7 +121,7 @@ function @zeta:3rd:is-vendor-pkg() {
 function @zeta:3rd:get-vendor-path() {
   local _path_; _path_="$(command -v $1)"
   [[ $? -ne 0 ]] && return
-  _path_="$(realpath "${_path_}")"; _path_="${_path_%/*}"
+  _path_="$(realpath -eq "${_path_}")"; _path_="${_path_%/*}"
   @zeta:3rd:is-vendor-pkg "${_path_}" && echo "${_path_}"
 }
 
@@ -136,8 +136,8 @@ function @zeta:3rd:get-pkg-version() {
 function @zeta:3rd:create-links() {
   local sym="${ZETA_DIR}/3rd/pick/$1"
   local app="${ZETA_DIR}/3rd/vendor/$2"
-  [[ -h "${app}" ]] && app="$(realpath "${app}")"
-  if [[ -f "${app}" ]]; then
+  [[ -h "${app}" ]] && app="$(realpath -eq "${app}")"
+  if [[ -n "${app}" && -f "${app}" ]]; then
     ln -sTf "${app}" "${sym}" # -s 符号链接 -f 若已存在则删除后重建
     printf "Create $(@D9 '3rd/pick/')$(@G3 "%-12s") " "$1"
     echo "$(@D9 '->') $(@Y3 "${app}")"
@@ -159,7 +159,7 @@ function @zeta:3rd:delete-links() {
   for binEXE in $(ls "${pkgBIN}"); do
     [[ -h "${pickDIR}/${binEXE}" ]] && {
       printf "Delete $(@D9 '3rd/pick/')$(@Y3 %-12s) $(@D9 '->') " ${binEXE}
-      echo "$(@G3 "$(realpath "${pickDIR}/${binEXE}")")"
+      echo "$(@G3 "$(realpath -eq "${pickDIR}/${binEXE}")")"
       rm -f "${pickDIR}/${binEXE}"
     }
   done
@@ -180,9 +180,32 @@ function @zeta:3rd:zeta-switch() {
   done
 }
 
+function @zeta:3rd:update-tools() {
+  local dst="${ZETA_DIR}/3rd/tools"  mid  src  fn  abs
+  if [[ -n "$(command ls "${dst}")" ]]; then
+    for fn in $(command ls "${dst}"); do
+      abs="$(realpath -eq "${dst}/${fn}")"
+      [[ $? -ne 0 || -z "${abs}" ]] && rm -f "${dst}/${fn}"
+    done
+  fi
+  for mid in gems rust/cargo; do
+    src="${ZETA_DIR}/3rd/vendor/${mid}/bin"
+    [[ -z "$(command ls "${src}")" ]] && continue
+    for fn in $(command ls "${src}"); do
+      abs="$(realpath -eq "${src}/${fn}")"
+      [[ $? -ne 0 || -z "${abs}" ]] && continue
+      ln -sTf  "${abs}"  "${dst}/${fn}"
+      echo "Create $(@D9 3rd/tools/)$(@G3 ${fn}) $(@D9 '->') $(@Y3 "${abs}")"
+    done
+  done
+}
+
 function @zeta:3rd:usage-help() {
   local _PKGS_=( cmake  java  node  rust  go  nim ) app version
-  echo; echo "-> $(@D9 zeta-switch) $(@R3 reset) $(@G3 PKG)"
+  echo
+  echo "-> $(@D9 zeta-switch) $(@R3 reset)   $(@G3 PKG)"
+  echo "-> $(@D9 zeta-switch) $(@R3 update-tools)"
+  echo
   for app in ${_PKGS_[@]}; do
     for version in $(@zeta:3rd:get-pkg-version ${app}); do
       printf -v app "%-5s" "${app}"
@@ -194,12 +217,15 @@ function @zeta:3rd:usage-help() {
 
 function zeta-switch() {
   [[ $# -eq 0 || -z "$1" ]] && { @zeta:3rd:usage-help; return; }
-  [[ "$1" == reset ]] && {
+
+  if [[ "$1" == reset ]]; then
     case "$2" in
       cmake|java|node|rust|go|nim) @zeta:3rd:delete-links $2; return ;;
       *) echo "invalid $(@D9 3rd/vendor/)$(@R3 $2) package"; return 1 ;;
     esac
-  }
+  elif [[ "$1" == update-tools ]]; then
+    @zeta:3rd:update-tools; return
+  fi
 
   case "$1" in
     cmake) @zeta:3rd:zeta-switch cmake "$2" ;; # cmake/$2/bin/*
@@ -222,9 +248,9 @@ if [[ -n "${ZSH_VERSION:-}" ]]; then
   function @zeta:comp:zeta-switch() {
     local -A opt_args
     local context state state_descr line
-    local -a _apps_=( cmake  java  node  rust  go  nim )
+    local -a _apps_=( cmake  java  node  rust  go  nim ) _keys_
     function @zeta:once±zcomps() {
-      local -a _keys_=( reset ${_apps_[@]} )
+      _keys_=( reset update-tools ${_apps_[@]} )
       _describe 'command' _keys_
     }
     # 语法 N:Message:Action 表示第 N 个参数执行 Action 行为
@@ -232,12 +258,12 @@ if [[ -n "${ZSH_VERSION:-}" ]]; then
     # NOTE Action 语法 ->XXX 表示将 state 设置为 XXX
     _arguments '1:what:@zeta:once±zcomps' '2:next:->todo'
     if [[ "todo" == "${state[1]}" ]]; then
-      if [[ "${line[1]}" != reset ]]; then
-        local -a _version_=( $(@zeta:3rd:get-pkg-version ${line[1]}) )
-        _describe 'command' _version_
-      else
-        _describe 'command' _apps_
-      fi
+      case "${line[1]}" in
+        update-tools) return ;;
+        reset) _describe 'command' _apps_; return ;;
+      esac
+      _keys_=( $(@zeta:3rd:get-pkg-version ${line[1]}) )
+      _describe 'command' _keys_
     fi
     unset -f @zeta:once±zcomps
   }
@@ -251,9 +277,15 @@ elif [[ -n "${BASH_VERSION:-}" ]]; then
     # 按空格分隔数组 COMP_WORDS, 最后输入词索引 COMP_CWORD
     [[ ${COMP_CWORD} -eq 3 ]] && return # 最多 2 个参数
     local -a _apps_=( cmake  java  node  rust  go  nim )
+
+    if [[ ${COMP_CWORD} -eq 1 ]]; then
+      COMPREPLY=( $(compgen -W "reset update-tools ${_apps_[*]}" -- "$2") )
+      return
+    fi
+
     case "$3" in
-      zeta-switch) COMPREPLY=( $(compgen -W "reset ${_apps_[*]}" -- "$2") ) ;;
-            reset) COMPREPLY=( $(compgen -W "${_apps_[*]}" -- "$2") ) ;;
+      update-tools) return ;;
+      reset) COMPREPLY=( $(compgen -W "${_apps_[*]}" -- "$2") ) ;;
       *) COMPREPLY=( $(compgen -W "$(@zeta:3rd:get-pkg-version "$3")" -- "$2") ) ;;
     esac
   }
